Move some stuff around
- RemoteFluidObjectHandle to runtime-utils
- Consolidate code in op-lifecycle/opSerialization.ts
