+++ packages/runtime/container-runtime/src/opLifecycle/index.ts
@@ -14,12 +14,15 @@ export {
 } from "./batchManager.js";
 export { BatchMessage, IBatch, IBatchCheckpoint, IChunkedOp } from "./definitions.js";
 export { DuplicateBatchDetector } from "./duplicateBatchDetector.js";
-export { Outbox, getLongStack, serializeOpContents } from "./outbox.js";
+export {
+	serializeOpContents,
+	ensureContentsDeserialized,
+} from "./opSerialization.js";
+export { Outbox, getLongStack } from "./outbox.js";
 export { OpCompressor } from "./opCompressor.js";
 export { OpDecompressor } from "./opDecompressor.js";
 export { OpSplitter, splitOp, isChunkedMessage } from "./opSplitter.js";
 export {
-	ensureContentsDeserialized,
 	InboundMessageResult,
 	BatchStartInfo,
 	RemoteMessageProcessor,
