+++ packages/runtime/container-runtime/src/opLifecycle/opContentsSerializer.ts
@@ -14,9 +14,8 @@ import {
 	isFluidHandle,
 	toFluidHandleInternal,
 	type ISerializedHandle,
+	RemoteFluidObjectHandle,
 } from "@fluidframework/runtime-utils/internal";
-
-import { RemoteFluidObjectHandle } from "../remoteObjectHandle.js";
 
 /**
  * @legacy
