+++ packages/runtime/runtime-utils/src/index.ts
@@ -26,6 +26,7 @@ export {
 	getNormalizedObjectStoragePathParts,
 	listBlobsAtTreePath,
 } from "./objectstorageutils.js";
+export { RemoteFluidObjectHandle } from "./remoteFluidObjectHandle.js";
 export { RequestParser } from "./requestParser.js";
 export { RuntimeFactoryHelper } from "./runtimeFactoryHelper.js";
 export {
@@ -45,7 +46,12 @@ export {
 	utf8ByteLength,
 } from "./summaryUtils.js";
 export { unpackChildNodesUsedRoutes } from "./unpackUsedRoutes.js";
-export { ReadAndParseBlob, seqFromTree, encodeCompactIdToString } from "./utils.js";
+export {
+	ReadAndParseBlob,
+	RuntimeHeaders,
+	seqFromTree,
+	encodeCompactIdToString,
+} from "./utils.js";
 export { isSnapshotFetchRequiredForLoadingGroupId } from "./snapshotUtils.js";
 export {
 	toDeltaManagerErased,
