+++ packages/runtime/runtime-utils/src/utils.ts
@@ -10,6 +10,21 @@ import {
 } from "@fluidframework/driver-definitions/internal";
 
 /**
+ * Accepted header keys for requests coming to the runtime.
+ * @internal
+ */
+export enum RuntimeHeaders {
+	/**
+	 * True to wait for a data store to be created and loaded before returning it.
+	 */
+	wait = "wait",
+	/**
+	 * True if the request is coming from an IFluidHandle.
+	 */
+	viaHandle = "viaHandle",
+}
+
+/**
  * Reads a blob from storage and parses it from JSON.
  *
  * @internal
